parser/requirements.txt
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
boto3==1.36.26
botocore==1.36.26
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.17.0
frozendict==2.4.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
multitasking==0.0.11
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parsel==1.10.0
peewee==3.17.9
platformdirs==4.3.6
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
s3transfer==0.11.2
Scrapy==2.12.0
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
tldextract==5.1.3
Twisted==24.11.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
w3lib==2.3.1
yfinance==0.2.54
zope.interface==7.2
